requirements.txt
transformers==4.15.0
#torch==1.11.0
#torchvision==0.12.0
#torchaudio==0.11.0
einops==0.4.0
torchmetrics==0.7.0
tntorch==1.0.1
pytorch-lightning==1.3.5